fix: Remove unused import and unnecessary cast
- Remove unused BCRYPT_COST import from auth/password.rs
- Fix unnecessary cast (usize -> usize) in auth/user_store.rs
- Keep JWT_EXPIRATION_SECS imports (used in tests)

+++ src/infrastructure/auth/user_store.rs
@@ -15,7 +15,7 @@ use std::path::Path;
 use std::time::{SystemTime, UNIX_EPOCH};
 
 /// Minimum JWT secret length for security (alias for centralized constant)
-pub const MIN_JWT_SECRET_LENGTH: usize = JWT_SECRET_MIN_LENGTH_STRICT as usize;
+pub const MIN_JWT_SECRET_LENGTH: usize = JWT_SECRET_MIN_LENGTH_STRICT;
 
 /// Characters used for password generation (alphanumeric + special)
 const PASSWORD_CHARSET: &[u8] =
